rename: theme에 있던 HeaderLogo 다른 파일로 이동
