+++ lib/header.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class Header extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    double statusBarHeight = MediaQuery.of(context).padding.top;
+
+    return Positioned(
+      left: 9,
+      top: statusBarHeight + 10,
+      child: SizedBox(
+        width: 150,
+        height: 35,
+        child: Image.asset(
+          'assets/icons/logo.png',
+          fit: BoxFit.contain,
+        ),
+      ),
+    );
+  }
+}
