+++ lib/theme.dart
@@ -15,26 +15,6 @@ class Background extends StatelessWidget {
   }
 }
 
-class HeaderLogo extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    double statusBarHeight = MediaQuery.of(context).padding.top;
-
-    return Positioned(
-      left: 9,
-      top: statusBarHeight + 10,
-      child: SizedBox(
-        width: 150,
-        height: 35,
-        child: Image.asset(
-          'assets/icons/logo.png',
-          fit: BoxFit.contain,
-        ),
-      ),
-    );
-  }
-}
-
 class font {
   static ThemeData lightTheme = ThemeData(
     fontFamily: 'Pretendard',
